pyproject.toml
[project]
name = "dataset"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.5.0",
    "huggingface-hub>=1.4.0",
    "ijson>=3.4.0.post0",
    "openai>=2.17.0",
    "pandas>=3.0.0",
    "pyarrow>=23.0.0",
    "python-dotenv>=1.2.1",
    "tqdm>=4.67.3",
]
